requirements.txt
opencv-python
ultralytics
PyYAML
torch
gTTS
pygame
numpy